requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.7.0
bcrypt==4.2.1
cachetools==5.5.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorlog==6.9.0
fastapi==0.115.7
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
greenlet==3.1.1
h11==0.14.0
httplib2==0.22.0
idna==3.10
oauth2client==4.1.3
oauthlib==3.2.2
proto-plus==1.26.1
protobuf==6.30.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.2.1
python-decouple==3.8
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
sqlmodel==0.0.22
starlette==0.41.3
typing_extensions==4.12.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
